Engine and Content Updates
Edited textures, added new blocks (Snowy Grass and Ice), modified player
damage, fixed collision.
Also began to work on terrain generation within chunks.
Made ore generation more realistic.
Made block rendering and culling more efficient (culls before looping
through list of blocks)

+++ src/net/adventuria/level/io/LevelSaver.java
@@ -19,7 +19,7 @@ public class LevelSaver
 	{
 		List<Object> saveData = new ArrayList<>();
 		
-		saveData.add("chrpos: {" + Integer.toHexString(Component.world.getPlayer().getBlockX()) + ", " + Integer.toHexString(Component.world.getPlayer().getBlockY()) + "}" + DATA_SEPARATOR);
+		saveData.add("chrpos: {" + Double.toHexString(Component.world.getPlayer().getBlockX()) + ", " + Double.toHexString(Component.world.getPlayer().getBlockY()) + "}" + DATA_SEPARATOR);
 		saveData.add("crntchnk: {");
 		
 		for(Block[] bArray : w.getCurrentChunk().getBlocks())
